requirements.txt
# --- Core backend ---
fastapi[standard]
uvicorn

# --- OpenAI & LLM ---
openai
tiktoken
python-dotenv

# --- LangGraph & LangChain ---
langgraph
langchain

pymongo
